Screenshot: Review1
https://github.com/xamarin/Essentials/pull/1433#pullrequestreview-505007197
