+++ Xamarin.Essentials/Platform/Platform.macos.cs
@@ -1,7 +1,10 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Runtime.InteropServices;
 using AppKit;
 using CoreFoundation;
+using CoreGraphics;
 using Foundation;
 using ObjCRuntime;
 
@@ -17,6 +20,46 @@ namespace Xamarin.Essentials
                 throw new InvalidOperationException("Could not find current window.");
 
             return window;
+        }
+    }
+
+    internal static class ApplicationServices
+    {
+        const string appServicesPath = "/System/Library/Frameworks/ApplicationServices.framework/Versions/Current/ApplicationServices";
+
+        // https://developer.apple.com/documentation/coregraphics/1454595-cgdisplaycreateimage
+        [DllImport(appServicesPath, EntryPoint = "CGDisplayCreateImage")]
+        static extern /* CGImageRef */ IntPtr CGDisplayCreateImage(int displayId);
+
+        internal static CGImage GetScreenshot()
+        {
+            using var pool = new NSAutoreleasePool();
+
+            // Get location from current window,
+            // if there is no window (e.g. menu bar app) we use the mouse location
+            CGPoint location;
+            var window = Platform.GetCurrentWindow(false);
+            if (window != null)
+            {
+                location = window.Frame.Location;
+            }
+            else
+            {
+                location = NSEvent.CurrentMouseLocation;
+            }
+
+            var screens = new List<NSScreen>(NSScreen.Screens);
+            var screen = screens.FirstOrDefault(obj => obj.Frame.Contains(location));
+            if (screen == null)
+            {
+                throw new InvalidOperationException("No screen found");
+            }
+
+            var windowNumber = (NSNumber)screen.DeviceDescription["NSScreenNumber"];
+            var displayId = windowNumber.Int32Value;
+            var handle = CGDisplayCreateImage(displayId); // CGWindowListCreateImage not working
+            var image = new CGImage(handle);
+            return image;
         }
     }
